Add animation for detail view in tablet-portrait mode
Added an animation showing the expansion of the ViewHolder when user
selects a recipe while in tablet-portrait mode. In tablet-portrait mode
the ViewHolder inflates a fragment showing the recipe details to make
better use of the space.

The animation uses a new class RecipeItemAnimator to play the animation.
This class will be expanded to encompass all animations used by
AdapterRecipe.

A new layout manager StaggeredGridLayoutManagerWithSmoothScroll was
created to allow smooth scroll animation to the top of the ViewHolder
upon the end of the scale animation.

There are still many bugs to work out:
 - Occasionally an invalid view will show behind the rest of the items
   attached to the recycler view
 - Clicking on the invalid view will cause the entire app to crash with
   java.lang.ArrayIndexOutOfBoundsException error
 - Fragment will not inflate properly occasionally and leave a blank
   view showing. Sometimes clicking it will make it properly inflate,
   other times it will crash.