Improve performance of shared element transitions
Previously, getSupportPostponeEnterTransition was never called due to
incorrect query of the database to see if the recipe had been previously
loaded. Fixed the error and set the fragment to load the image from
source if it exists.

A background was also added to the layout of
R.layout.fragment_recipe_details on details_recipe_content1 & 2 so the
slide animation would have a background instead of just text.

Also fixed error where FragmentRecipeDetails would not load details of
an imported recipe due to the CursorLoader pointing to an incorrect URI.