Refactor RecipeListAsyncTasks as Services
App will be converted to run the Service in the background using
GCMTaskService throughout the day instead of when the app is run. This
will work to prevent data from getting stale and show a more fluid UI
as the newest recipes will already be loaded upon app start instead of
being updated while the user is looking at the app.

Fixed a small bug where recipes imported from Epicurious utilizing
placeholder images, but having actual an actual image when the details
are loaded would cause an infinite loop as a new recipe is generated
that doesn't match the recipeSourceId loaded into ActivityRecipeDetails.