Add capability for app to automatic sync on network state change
If the app hasn't been synced in over six hours and the previous network
state was disconnected, the app will automatically initiate a sync when
the network reconnects to an active network.

The app will also initiate a sync if it detects a sync hasn't occurred
in over six hours or if there are no recipes in the database.