Optimize image loading in ActivityRecipeDetails
Previously, the image URL was obtained by querying the database for the
URL, this meant that when searching for recipes when no entries are
created for the recipes in the database, no image was ever loaded into
the ImageView. This prevented the window transition from initiating and
loading the rest of the recipe information into view.