Fix bug preventing created recipes from showing in ActivityMyRecipes
Newly created recipes were not stored with an asterisk prior to the
recipe ID, so it wasn't being properly filtered in ActivityMyRecipes.

Also fixed bug when trying to create a recipe with a photo that is too
large. Bitmap is now properly resized prior to saving so it can be
loaded in FragmentRecipeDetails.