Add shopping list Dialog for adding ingredients to shopping list
Allow the user to add items to an internal shopping list (to be made) by
displaying a new ShoppingListDialog that includes all ingredients for a
given recipe. A CheckBox is included to allow the user to selectively
add ingredients to the shopping list.

Added method Utilities#removePreparation that will attempt to strip
preparation verbiage from the ingredient to display a cleaner shopping
list.

Increase the ingredient key line dimension to 72dp to better allow for
awkward quantities such as 1 (10.5 oz) can.