Update animations in ActivityShoppingList
FAB now only shows when there are items that have been checked off by
the user. Additionally, a ScrollListener has been added to mRecyclerView
so that when the user is scrolling the list, the FAB auto-hides itself
if there are items already checked off.

Clicking the FAB to delete items from the shopping list will correctly
animate their removal. By utilizing setHasStableIds on the
AdapterIngredient and setting the item IDs to the primary key, the
Adapter is informed which items are removed.

Added a light grey background to the checked items so that they are more
easily identified by the user.