Update layout and functionality of AddRecipeDialog
Previously, the dialog merely showed a list of recipes to the user,
making it difficult to select a specific recipe they had in mind.
Options for filtering by favorites, custom-recipes, and searching by
name were added to the layout to make it easier to narrow down recipes
the user would like to add to the recipe book.