Add option to modify a generic recipe when first opened
When the user first imports a generic recipe, it is loaded into
ActivityRecipeDetails. However, because the algorithm is not perfect,
there will often be times where the description is too long, the
ingredients are incorrect, etc. A Snackbar was added allowing the user
the option to edit the detail from ActivityCreateDetails with the
difference that the original is deleted from the list after they finish
editing it so that there is no copy with a mistake saved.

It is not saved as the original because it is impossible to know what
changes the user is making. If they change even one ingredient or one
direction, then it is no longer the original and should not be saved as
such.