Update ActivityRecipeDetails
Previous changes to the Activity altered FragmentActivityDetails so that
the ImageView was moved to the Activity in the phone layout. Therefore,
the ImageView for tablets was never populated when called from a master-
flow layout. The loading of images was re-implemented into the Fragment
for tablets.

Refactored a few methods for easier editing and reading of the code.